refactor: move FundingRoundOverviewCards component to it's own file
